rename: 트리 등록 폼 컴포넌트 폴더 구조 수정
